Change codebook to use a trie
This change hugely improves the speed of the compressor without any
change to the result. For example, the previous hash map implementation
took about 60s to process 14 MB or text and the new implementation takes
about 2s. This was primarily due to the hash map being called on many
strings that are not in the map. The use of a trie eliminates this as we
can test if the prefix is not present in the map at any point.